Moved code in lambda.ts to combiators.ts
