+++ package.json
@@ -1,11 +1,12 @@
 {
-	"name": "@hypothesize/standard",
+	"name": "@agyemanjp/standard",
 	"version": "1.0.0",
 	"description": "Standard library for JavaScript/TypeScript",
-	"main": "dist/index.js",
-	"types": "dist/index.d.ts",
+	"main": "dist/_index.js",
+	"types": "dist/_index.d.ts",
 	"files": [
-		"dist",
+		"dist/*.js",
+		"dist/*.d.ts",
 		"!/dist/**/*test.*"
 	],
 	"scripts": {
